requirements.txt
requests>=2.31.0
openpyxl>=3.1.2
python-dotenv>=1.0.0
google-genai>=1.0.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
PyPDF2>=3.0.0
python-docx>=1.0.0
